httpx/requirements.txt
fastapi[all]
httpx
aiofile